Add "Professor" class
Add a "Professor" class, which provides the same functionality as Scientist, but as an instance-based class, rather than a static class.
This is to make activation via dependency injection easier by removing the need to set a shared static property with the desired IResultPublisher instance that might require activation via an IoC container.